// fixtures/sql/generated/valid/nested/path_mapping.ts
// @generated by sqlcomp. Do not edit.

export type nestedPathMapping_Input = Record<string, never>;

export type nestedPathMapping_Row = {
  parentBigintNnCol: string;
  childVarchar32NnCol: string;
};

export type nestedPathMapping_Output = nestedPathMapping_Row[];

export function nestedPathMapping(
  _input: nestedPathMapping_Input = {},
): { sql: string; params: readonly [] } {
  return {
    sql: "\nSELECT\n  p.bigint_nn_col AS parentBigintNnCol,\n  c.varchar_32_nn_col AS childVarchar32NnCol\nFROM fixture_all_column_type AS p\nINNER JOIN fixture_child AS c\n  ON c.parent_bigint_nn_col = p.bigint_nn_col\nWHERE c.varchar_32_nn_col = 'child-a';\n",
    params: [] as const,
  };
}